List borrows + start to improved search Not yet any way to borrow a book though.
The passed string should be matched against a separate list of authors
and their books in case the user supplied an author. A year expression
should also exist as an option. The question is if the books also should
be stored in a B-tree along side the Bk-tree in order to retrieve books
by year expression or if a filter should just be applied in the end to
check if the book was published withing the range provided.