Containerization using docker compose
Two containers, frontend and backend run along side a Caddy reverse
proxy that maps / to the frontend and /backend to the backend. This
ensures they are on the same origin and we can use SameSite cookies.

https is not setup yet as this will only be run locally over LAN